fix: `go test -f` needs backslash escaping on Windows (#277)
This adds support for testing on Windows with plenary. I was able to get Busted to run tests but it was a nightmare on Windows and I gave up. The entire setup also involved a lot of moving parts, so I feel PlenaryBusted is still the solution I'd like to keep.