Add additional possible error cause
To get faster feedback, I normally use `-vet="off"` in a shell. Placing that in the list of args cased the the "no tests" error. But passing them as `-vet=off` did the trick - so I'd just share the experience :)

I guess it's because it's the shell itself that interprets the command line, that handles the `"`s, so `go test` sees the value without `"`, so I just added that explanation for context.