feat: introduce streaming test result processing (#373)
Streaming allows for quicker, near-realtime, feedback on individual test results
when running a large test suite.

The intent is to keep the streaming a "hot path" with the primary goal of
reporting back the status and output of the test that executed. Therefore, some
work will be deferred until "finalization", which is when the test command has
finished, so to avoid O(n^2) operations which would cause the streaming to
appear as sluggish/choppy and even take longer time than the test execution
itself.

- Add streaming capability. Total rewrite of all `go test` (or `gotestsum`)
  output processing:
  - Proper test name collection; if a test is executed, neotest-golang will know
    about it. If AST test detection did not discover the test, the status/output
    does not get populated into the Neotest position. But if this test fails,
    neotest-golang will pick up on it and report failure.
  - Formal introduction of an asynchronous "hot path" streaming phase and a
    synchronous finalization phase.
  - Improved colorization of test output.
  - Inline hints (not only error diagnostics) via custom.
  - Integration tests!

Changes (non-breaking):

- Option `warn_test_not_executed` removed: This is more difficult to detect now,
  and somewhat unclear what it really means. Might be something that can be
  re-introduced in the future. There are other reports in the log which will
  appear instead and might be more helpful:
  - "Test was executed but not detected": when tests were executed, but were not
    detected during the neotest-golang test detection phase.
  - "Test results not populated for the following Neotest positions:": when a
    Neotest position exists, but for some reason the test status/output was not
    registered during test output/result processing.

As a side-effect, this might open up (more than with the old output processing,
at least) for the potential ability to detect tests in runtime, that could not
be detected by AST parsing. The reason for this is we now will be able to detect
all tests names that gets executed by `go test`. However, I don't think Neotest
supports injecting positions into the test position tree after the AST test
detection phase, which can be a blocker for this.